move application tx handling to another module
